refactor(subagents): restore schema-derived types
